Chunk bit shift is faster

+++ src/kernel_lion.h
@@ -51,14 +51,14 @@
 #define DENSITY_LION_PREFERRED_EFFICIENCY_CHECK_CHUNKS              (1 << DENSITY_LION_PREFERRED_EFFICIENCY_CHECK_CHUNKS_SHIFT)
 
 #define DENSITY_LION_HASH32_MULTIPLIER                                  (uint32_t)0x9D6EF916lu
-#define DENSITY_LION_HASH16_MULTIPLIER                                  0xAD15lu
+#define DENSITY_LION_HASH16_MULTIPLIER                                  (uint32_t)0xAD15lu
 #define DENSITY_LION_CHUNK_HASH_BITS                                    16
 #define DENSITY_LION_BIGRAM_HASH_BITS                                   8
 
 #define DENSITY_LION_HASH_ALGORITHM(hash32, value32)                    hash32 = value32 * DENSITY_LION_HASH32_MULTIPLIER;\
                                                                         hash32 = (hash32 >> (32 - DENSITY_LION_CHUNK_HASH_BITS));
 
-#define DENSITY_LION_BIGRAM_HASH_ALGORITHM(bigram)                      (uint8_t) (((bigram) * DENSITY_LION_HASH16_MULTIPLIER) >> (16 - DENSITY_LION_BIGRAM_HASH_BITS))
+#define DENSITY_LION_BIGRAM_HASH_ALGORITHM(bigram)                      (uint8_t) ((bigram * DENSITY_LION_HASH16_MULTIPLIER) >> (16 - DENSITY_LION_BIGRAM_HASH_BITS))
 
 typedef enum {
     DENSITY_LION_FORM_CHUNK_PREDICTIONS,
